requirements.txt
Flask>=2.3.0
Flask-CORS>=4.0.0
python-dotenv>=0.19.0
